Separate vLLM notebook from the sglang one and add across-the-runs statistics.
